fix(osc.js): add explicit string typecheck for `receivedValue`
In other feedback callbacks, a conversion of a number is checked with `isNaN()`, but that doesn't apply here. This is a solution in the same vein but may not have the same effect.
- `String(null)` returns `"null"`
`String(undefined)` returns `"undefined"`
- `String({"someObjectKey":"some object's value"})` returns `"[object Object]"`

Alternative: make the check directly on `rx_args[0].value` before passing it into `String()`.